+++ ProjectFlightSchool/ProjectFlightSchool/Projectile.h
@@ -21,20 +21,20 @@ class Projectile
 		bool			mIsActive;
 		float			mLifeTime;
 		BoundingCircle*	mBoundingCircle;
-		int				mDamage;
+		float			mDamage;
 
 	public:
 		HRESULT			Update( float deltaTime );
 		HRESULT			Render();
 
-		void			SetDirection( unsigned int playerID, unsigned int id, XMFLOAT3 startPosition, XMFLOAT3 direction, float speed, float range, int damage );
+		void			SetDirection( unsigned int playerID, unsigned int id, XMFLOAT3 startPosition, XMFLOAT3 direction, float speed, float range, float damage );
 		void			SetIsActive( bool isActive );
 		bool			IsActive() const;
 		void			Reset();
 		BoundingCircle*	GetBoundingCircle() const;
 		unsigned int	GetPlayerID() const;
 		unsigned int	GetID() const;
-		unsigned int	GetDamage() const;
+		float			GetDamage() const;
 
 		HRESULT			Initialize();
 		void			Release();
